Fixes data directory ownership on upgrades
Addresses an issue where upgraded containers retain root ownership of files in the data directory if the previous version ran as root.

This change now checks for root-owned files and changes their ownership to the specified PUID:PGID if the container is not running as root (PUID!=0). This ensures that the user can access and modify their data after upgrading the container.